+++ app/src/main/java/xyz/flussikatz/searchmovie/data/ApiKey.kt
@@ -1,4 +1,4 @@
-package xyz.flussikatz.searchmovie.util
+package xyz.flussikatz.searchmovie.data
 
 object ApiKey {
     const val API_KEY = ""
